pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mylittletracker"
version = "0.1.0"
description = "Personal parcels tracking CLI that queries carrier APIs (e.g., Correos)."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
  { name = "Your Name" }
]
keywords = ["parcels", "tracking", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "python-dotenv>=1.0.0",
  "httpx>=0.25.0",
  "pydantic>=2.5.0",
  "beautifulsoup4>=4.12.0"
]

[project.urls]
Homepage = "https://example.com/mylittletracker"

[project.scripts]
mylittletracker = "mylittletracker.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

